LayoutTests/crypto/subtle/ec-import-jwk-key-export-jwk-key-public-expected.txt

Test importing a jwk public ECDH key and then export it in jwk format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS publicKey.kty is jwkKey.kty
PASS publicKey.crv is jwkKey.crv
PASS publicKey.x is jwkKey.x
PASS publicKey.y is jwkKey.y
PASS publicKey.key_ops is [ ]
PASS publicKey.ext is jwkKey.ext
PASS successfullyParsed is true

TEST COMPLETE
